requirements.txt
keras==2.12.0
scipy==1.10.1
shap==0.41.0
torch==1.10.1+cu113
tensorflow==2.12.0
transformers==4.24.0
biopython==1.78
numpy==1.22.3
setuptools==59.5.0